[FIX] gls picking and demo data
