+++ delivery_carrier_label_gls/__init__.py
@@ -2,5 +2,5 @@ from . import company
 from . import config
 from . import delivery
 from . import report
-from . import stk2
+from . import picking
 from . import stock
