+++ delivery_carrier_label_gls/demo/product.xml
@@ -42,15 +42,5 @@
             <field name="location_id" ref="stock.stock_location_stock"/>
         </record>
 
-        <function model="stock.inventory" name="action_confirm">
-            <function eval="[[('id', '=', ref('stock_inventory_carrier'))]]"
-                      model="stock.inventory" name="search"/>
-        </function>
-
-        <function model="stock.inventory" name="action_done">
-            <function eval="[[('id', '=', ref('stock_inventory_carrier'))]]"
-                      model="stock.inventory" name="search"/>
-        </function>
-
     </data>
 </openerp>
