+++ delivery_carrier_label_gls/stock_view.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+  <data>
+
+<!-- From DELIVERY module on STOCK.PICKING -->
+<!-- FORM -->
+<record id="view_picking_withcarrier_out_form" model="ir.ui.view">
+  <field name="model">stock.picking</field>
+  <field name="inherit_id" ref="delivery.view_picking_withcarrier_out_form"/>
+  <field name="arch" type="xml">
+    <xpath expr="//field[@name='pack_operation_ids']/tree/field[@name='result_package_id']"
+           position="after">
+        <button type="object" name="open_tracking_url"
+            class="oe_button fa_truck"
+            attrs="{'invisible': [('result_package_id', '=', False)]}"
+            icon="terp-accessories-archiver+"
+            />
+                <!--class="glyphicon glyphicon-qrcode"-->
+                <!--icon="fa_truck"-->
+                <!--icon="terp-accessories-archiver+"-->
+    </xpath>
+  </field>
+</record>
+
+<record id="view_quant_package_form" model="ir.ui.view">
+    <field name="model">stock.quant.package</field>
+    <field name="inherit_id" ref="stock.view_quant_package_form"/>
+    <field name="arch" type="xml">
+        <button name="unpack" position="after">
+            <button type="object" name="open_tracking_url"
+                    string="Carrier Tracking"
+                    attrs="{'invisible': [('parcel_tracking', '=', False)]}"
+                    help="Track your parcel in carrier information system"/>
+        </button>
+    </field>
+</record>
+
+  </data>
+</openerp>
